+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## June 2024 (by David C. Stauffer)
+
+* Drop support for Python v3.9, and use newer typing features like dict, list, tuple and | for Union
+
 ## August 2022 (by David C. Stauffer)
 
 * Split into a stand-alone library using the latest standards of pyproject.toml, plus single source versioning, generic named src folder, separated tests and robust testing with pytest, mypy, black and coverage.